Enhance UI consistency with SVG icons and fix bilingual translations
- Replace microphone emoji with elegant SVG icon in SpeechTextarea help text
- Replace company/department/manager badge emojis with professional SVG icons in evaluate page
- Fix bilingual text translations using correct keys (t.dashboard.createdBy, t.dashboard.deadlineSetBy)
- Fix textarea focus outline alignment with microphone button positioning
- Improve visual consistency across evaluation interface components
- Maintain proper flexbox alignment and spacing for all icon-text combinations
